RF+TST: refactor dataobj attribute into own class
Add interface to dataobj to an abstract class for easier inheritance by
others (particularly CIFTI2).  Make SpatialImage inherit from this
class.  Add tests for filebasedimages and dataobj_images.
