+++ nibabel/tests/test_filebasedimages.py
@@ -1,0 +1,74 @@
+""" Testing filebasedimages module
+"""
+
+from itertools import product
+
+import numpy as np
+
+from nibabel.filebasedimages import FileBasedHeader, FileBasedImage
+
+from nibabel.tests.test_image_api import GenericImageAPI
+
+
+class FBNumpyImage(FileBasedImage):
+    header_class = FileBasedHeader
+    valid_exts = ('.npy',)
+    files_types = (('image', '.npy'),)
+
+    def __init__(self, arr, header=None, extra=None, file_map=None):
+        super(FBNumpyImage, self).__init__(header, extra, file_map)
+        self.arr = arr
+
+    @property
+    def shape(self):
+        return self.arr.shape
+
+    def get_data(self):
+        return self.arr
+
+    @classmethod
+    def from_file_map(klass, file_map):
+        with file_map['image'].get_prepare_fileobj('rb') as fobj:
+            arr = np.load(fobj)
+        return klass(arr)
+
+    def to_file_map(self, file_map=None):
+        file_map = self.file_map if file_map is None else file_map
+        with file_map['image'].get_prepare_fileobj('wb') as fobj:
+            np.save(fobj, self.arr)
+
+    def get_data_dtype(self):
+        return self.arr.dtype
+
+    def set_data_dtype(self, dtype):
+        self.arr = self.arr.astype(dtype)
+
+
+class TestFBImageAPI(GenericImageAPI):
+    """ Validation for FileBasedImage instances
+    """
+    # A callable returning an image from ``image_maker(data, header)``
+    image_maker = FBNumpyImage
+    # A callable returning a header from ``header_maker()``
+    header_maker = FileBasedHeader
+    # Example shapes for created images
+    example_shapes = ((2,), (2, 3), (2, 3, 4), (2, 3, 4, 5))
+    example_dtypes = (np.int8, np.uint16, np.int32, np.float32)
+    can_save = True
+    standard_extension = '.npy'
+
+    def obj_params(self):
+        # Create new images
+        def make_imaker(arr, header=None):
+            return lambda: self.image_maker(arr, header)
+
+        for shape, dtype in product(self.example_shapes, self.example_dtypes):
+            arr = np.arange(np.prod(shape), dtype=dtype).reshape(shape)
+            hdr = self.header_maker()
+            func = make_imaker(arr.copy(), hdr)
+            params = dict(
+                dtype=dtype,
+                data=arr,
+                shape=shape,
+                is_proxy=False)
+            yield func, params
